LPS-161236 Creates `components`, `hooks` folders and a new file called `utils` to organize files
