+++ modules/apps/frontend-js/frontend-js-walkthrough-web/src/main/resources/META-INF/resources/hooks/useLocalStorage.js
